refactor: components folder adapted
