+++ frontend/src/components/BaseModal/index.ts
@@ -1,0 +1,1 @@
+export * from './BaseModal';
