+++ frontend/src/components/Form/FormGrid/FormGrid.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+import { Box } from '@/components/Box';
+
+export const Row = styled(Box)`
+	display: flex;
+	align-items: center;
+`;
+
+export const Column = styled(Box)`
+	display: flex;
+	flex-direction: column;
+`;
