+++ frontend/src/components/Form/FormGrid/index.ts
@@ -1,13 +1,1 @@
-import styled from 'styled-components';
-
-import { Box } from '@/components/Box';
-
-export const Row = styled(Box)`
-	display: flex;
-	align-items: center;
-`;
-
-export const Column = styled(Box)`
-	display: flex;
-	flex-direction: column;
-`;
+export * from './FormGrid';
