+++ frontend/src/components/Form/PasswordInput/index.ts
@@ -1,0 +1,1 @@
+export * from './PasswordInput';
