+++ frontend/src/components/LayoutBox/LayoutBox.styles.ts
