+++ frontend/src/components/LayoutBox/index.ts
@@ -1,0 +1,1 @@
+export * from './LayoutBox';
