--- frontend/src/components/LayoutBox/index.tsx
@@ -1,30 +1,0 @@
-import { 
-	LayoutBoxContainer, 
-	LayoutBoxContent, 
-	LayoutBoxFooter, 
-	LayoutBoxFooterLeftSide, 
-	LayoutBoxFooterRightSide, 
-	LayoutBoxHeader, 
-	LayoutBoxHeaderButtonsContainer, 
-	LayoutBoxTitle
-} from "./styles";
-
-type LayoutBoxProps = {
-	children: React.ReactNode;
-}
-
-export function LayoutBox({ children }: LayoutBoxProps) {
-	return (
-		<LayoutBoxContainer>
-			{children}
-		</LayoutBoxContainer>
-	)
-}
-
-LayoutBox.Header = LayoutBoxHeader;
-LayoutBox.HeaderTitle = LayoutBoxTitle;
-LayoutBox.HeaderButtonsContainer = LayoutBoxHeaderButtonsContainer
-LayoutBox.Content = LayoutBoxContent;
-LayoutBox.Footer = LayoutBoxFooter;
-LayoutBox.FooterLeftSide = LayoutBoxFooterLeftSide;
-LayoutBox.FooterRightSide = LayoutBoxFooterRightSide;
