+++ frontend/src/components/NavLink/index.ts
@@ -1,0 +1,1 @@
+export * from './NavLink';
