+++ frontend/src/components/SEO/index.ts
@@ -1,0 +1,1 @@
+export * from './SEO';
