+++ frontend/src/components/SelectMonthAndYear/SelectMonthAndYear.styles.ts
