+++ frontend/src/components/SelectMonthAndYear/index.ts
@@ -1,0 +1,1 @@
+export * from './SelectMonthAndYear';
