+++ frontend/src/components/Table/Table.styles.ts
