+++ frontend/src/components/index.ts
@@ -1,0 +1,11 @@
+export * from './BaseModal';
+export * from './Box';
+export * from './Button';
+export * from './Header';
+export * from './LayoutBox';
+export * from './NavLink';
+export * from './SelectMonthAndYear';
+export * from './SEO';
+export * from './Spinner';
+export * from './Table';
+export * from './Form';
